Refactors CharacterStatsForm so it doesn't need props
