+++ src/components/characterCreation/CharacterStatsForm.tsx
@@ -1,53 +1,57 @@
-import React, { FC, useEffect, useState } from 'react';
-import { useQuery } from '@apollo/client';
+import React, { FC, useState } from 'react';
+import { useMutation, useQuery } from '@apollo/client';
+import { useHistory } from 'react-router-dom';
 import { Box } from 'grommet';
 
 import Spinner from '../Spinner';
 import { ButtonWS, HeadingWS } from '../../config/grommetConfig';
-import { PlaybookType } from '../../@types/enums';
-import { PlaybookCreator, StatsOption } from '../../@types/staticDataInterfaces';
 import PLAYBOOK_CREATOR, { PlaybookCreatorData, PlaybookCreatorVars } from '../../queries/playbookCreator';
+import SET_CHARACTER_STATS, { SetCharacterStatsData, SetCharacterStatsVars } from '../../mutations/setCharacterStats';
+import { CharacterCreationSteps } from '../../@types/enums';
+import { StatsOption } from '../../@types/staticDataInterfaces';
 import { useFonts } from '../../contexts/fontContext';
+import { useGame } from '../../contexts/gameContext';
 
-interface CharacterStatsFormProps {
-  playbookType: PlaybookType;
-  settingStats: boolean;
-  characterName: string;
-  handleSubmitStats: (statsOptionId: string) => void;
-  existingStatOption?: string;
-}
-
-const CharacterStatsForm: FC<CharacterStatsFormProps> = ({
-  playbookType,
-  settingStats,
-  characterName,
-  handleSubmitStats,
-  existingStatOption,
-}) => {
-  const [selectedStatsOption, setSelectedStatsOption] = useState<StatsOption | undefined>();
-  const [pbCreator, setPbCreator] = useState<PlaybookCreator | undefined>();
-
+const CharacterStatsForm: FC = () => {
+  // ------------------------------------------------------- Hooks --------------------------------------------------------- //
+  const { game, character, userGameRole } = useGame();
   const { crustReady } = useFonts();
 
-  const { data: pbCreatorData, loading: loadingPbCreator } = useQuery<PlaybookCreatorData, PlaybookCreatorVars>(
+  // -------------------------------------------------- Component state ---------------------------------------------------- //
+  const existingStatsOptionId = character?.statsBlock?.statsOptionId;
+  const [selectedStatsOption, setSelectedStatsOption] = useState<StatsOption | undefined>();
+
+  // --------------------------------------------------3rd party hooks ----------------------------------------------------- //
+  const history = useHistory();
+
+  // -------------------------------------------------- Graphql hooks ---------------------------------------------------- //
+  const { data: pbCreatorData } = useQuery<PlaybookCreatorData, PlaybookCreatorVars>(
     PLAYBOOK_CREATOR,
-    {
-      variables: { playbookType },
-    }
+    // @ts-ignore
+    { variables: { playbookType: character?.playbook }, skip: !character?.playbook }
+  );
+  const statsOptions = pbCreatorData?.playbookCreator.statsOptions;
+  const [setCharacterStats, { loading: settingStats }] = useMutation<SetCharacterStatsData, SetCharacterStatsVars>(
+    SET_CHARACTER_STATS
   );
 
-  useEffect(() => {
-    !!pbCreatorData && setPbCreator(pbCreatorData.playbookCreator);
-  }, [pbCreatorData]);
+  // ---------------------------------------- Component functions and variables ------------------------------------------ //
+  const handleSubmitStats = async (statsOptionId: string) => {
+    if (!!userGameRole && !!character && !!game) {
+      try {
+        await setCharacterStats({
+          variables: { gameRoleId: userGameRole.id, characterId: character.id, statsOptionId },
+          // refetchQueries: [{ query: GAME, variables: { gameId } }],
+        });
+        history.push(`/character-creation/${game.id}?step=${CharacterCreationSteps.selectGear}`);
+        window.scrollTo({ top: 0, left: 0, behavior: 'smooth' });
+      } catch (error) {
+        console.error(error);
+      }
+    }
+  };
 
-  if (loadingPbCreator || !pbCreatorData || !pbCreator) {
-    return (
-      <Box fill background="transparent" justify="center" align="center">
-        <Spinner />
-      </Box>
-    );
-  }
-
+  // -------------------------------------------------- Render component  ---------------------------------------------------- //
   return (
     <Box
       fill
@@ -58,61 +62,60 @@ const CharacterStatsForm: FC<CharacterStatsFormProps> = ({
       justify="start"
     >
       <Box width="50vw" flex="grow">
-        <HeadingWS
-          level={2}
-          crustReady={crustReady}
-          textAlign="center"
-        >{`WHAT ARE ${characterName.toUpperCase()}'S STRENGTHS AND WEAKNESSES?`}</HeadingWS>
+        <HeadingWS level={2} crustReady={crustReady} textAlign="center">{`WHAT ARE ${
+          !!character?.name ? character.name.toUpperCase() : '...'
+        }'S STRENGTHS AND WEAKNESSES?`}</HeadingWS>
         <HeadingWS level={4} textAlign="start">
           Choose a set:
         </HeadingWS>
         <Box fill margin={{ bottom: '48px' }}>
-          {pbCreator.statsOptions.map((opt) => (
-            <Box
-              data-testid={`${opt.id}-stats-option-box`}
-              key={opt.id}
-              direction="row"
-              justify="around"
-              align="center"
-              border={opt.id === existingStatOption}
-              background={{ color: 'neutral-1', opacity: opt.id === existingStatOption ? 0.5 : 0 }}
-              hoverIndicator={{ color: 'neutral-1', opacity: 0.4 }}
-              onClick={() => setSelectedStatsOption(opt)}
-              gap="6px"
-              style={{ minHeight: '52px' }}
-            >
-              <Box direction="row" align="center" gap="12px">
-                <HeadingWS level={4}>COOL:</HeadingWS>
-                <HeadingWS crustReady={crustReady} color="brand" level={3}>
-                  {opt.COOL}
-                </HeadingWS>
+          {!!statsOptions &&
+            statsOptions.map((opt) => (
+              <Box
+                data-testid={`${opt.id}-stats-option-box`}
+                key={opt.id}
+                direction="row"
+                justify="around"
+                align="center"
+                border={opt.id === existingStatsOptionId}
+                background={{ color: 'neutral-1', opacity: opt.id === existingStatsOptionId ? 0.5 : 0 }}
+                hoverIndicator={{ color: 'neutral-1', opacity: 0.4 }}
+                onClick={() => setSelectedStatsOption(opt)}
+                gap="6px"
+                style={{ minHeight: '52px' }}
+              >
+                <Box direction="row" align="center" gap="12px">
+                  <HeadingWS level={4}>COOL:</HeadingWS>
+                  <HeadingWS crustReady={crustReady} color="brand" level={3}>
+                    {opt.COOL}
+                  </HeadingWS>
+                </Box>
+                <Box direction="row" align="center" gap="12px">
+                  <HeadingWS level={4}>HARD:</HeadingWS>
+                  <HeadingWS crustReady={crustReady} color="brand" level={3}>
+                    {opt.HARD}
+                  </HeadingWS>
+                </Box>
+                <Box direction="row" align="center" gap="12px">
+                  <HeadingWS level={4}>HOT:</HeadingWS>
+                  <HeadingWS crustReady={crustReady} color="brand" level={3}>
+                    {opt.HOT}
+                  </HeadingWS>
+                </Box>
+                <Box direction="row" align="center" gap="12px">
+                  <HeadingWS level={4}>SHARP:</HeadingWS>
+                  <HeadingWS crustReady={crustReady} color="brand" level={3}>
+                    {opt.SHARP}
+                  </HeadingWS>
+                </Box>
+                <Box direction="row" align="center" gap="12px">
+                  <HeadingWS level={4}>WEIRD:</HeadingWS>
+                  <HeadingWS crustReady={crustReady} color="brand" level={3}>
+                    {opt.WEIRD}
+                  </HeadingWS>
+                </Box>
               </Box>
-              <Box direction="row" align="center" gap="12px">
-                <HeadingWS level={4}>HARD:</HeadingWS>
-                <HeadingWS crustReady={crustReady} color="brand" level={3}>
-                  {opt.HARD}
-                </HeadingWS>
-              </Box>
-              <Box direction="row" align="center" gap="12px">
-                <HeadingWS level={4}>HOT:</HeadingWS>
-                <HeadingWS crustReady={crustReady} color="brand" level={3}>
-                  {opt.HOT}
-                </HeadingWS>
-              </Box>
-              <Box direction="row" align="center" gap="12px">
-                <HeadingWS level={4}>SHARP:</HeadingWS>
-                <HeadingWS crustReady={crustReady} color="brand" level={3}>
-                  {opt.SHARP}
-                </HeadingWS>
-              </Box>
-              <Box direction="row" align="center" gap="12px">
-                <HeadingWS level={4}>WEIRD:</HeadingWS>
-                <HeadingWS crustReady={crustReady} color="brand" level={3}>
-                  {opt.WEIRD}
-                </HeadingWS>
-              </Box>
-            </Box>
-          ))}
+            ))}
           <Box direction="row" justify="end" width="50vw" margin={{ top: '12px' }}>
             <ButtonWS
               primary
